Create Entity

+++ StoreProject.Application/DTOs/Coupon/CreateCouponDto.cs
@@ -9,7 +9,7 @@ namespace StoreProject.Application.DTOs.Coupon
     public class CreateCouponDto
     {
         public string CouponCode { get; set; }
-        public decimal MinAmount { get; set; }
-        public decimal DiscountAmount { get; set; }
+        public double MinAmount { get; set; }
+        public double DiscountAmount { get; set; }
     }
 }
