+++ StoreProject.Application/DTOs/Coupon/UpdateCouponDto.cs
@@ -10,7 +10,7 @@ namespace StoreProject.Application.DTOs.Coupon
     public class UpdateCouponDto : BaseDto
     {
         public string CouponCode { get; set; }
-        public decimal MinAmount { get; set; }
-        public decimal DiscountAmount { get; set; }
+        public double MinAmount { get; set; }
+        public double DiscountAmount { get; set; }
     }
 }
